agent/requirements.txt
google-genai
mcp
openai
python-dotenv